+++ exchangelib/queryset.py
@@ -111,12 +111,16 @@ class QuerySet(object):
             return self.folder.find_items(self.q, **find_item_kwargs)
         if complex_fields_requested:
             # The FindItems service does not support complex field types. Fallback to getting ids and calling GetItems
-            ids = self.folder.find_items(self.q, **find_item_kwargs)
-            items = self.folder.fetch(ids=ids, only_fields=additional_fields)
+            # TODO: This is greedy because thread_pool.map() is greedy
+            items = self.folder.fetch(
+                ids=self.folder.find_items(self.q, **find_item_kwargs),
+                only_fields=additional_fields
+            )
         else:
             find_item_kwargs['additional_fields'] = additional_fields
-            items = self.folder.find_items(self.q, find_item_kwargs)
+            items = self.folder.find_items(self.q, **find_item_kwargs)
         if self.order_fields:
+            # Ordering and reversing is greedy
             assert isinstance(self.order_fields, tuple)
             # Sorting in Python is stable, so when we search on multiple fields, we can do a sort on each of the
             # requested fields in reverse order. Reverse each sort operation if the field is prefixed with '-'
@@ -131,25 +135,36 @@ class QuerySet(object):
                         setattr(i, f, None)
                     return i
 
-                items = map(clean_item, items)
+                return (clean_item(i) for i in items)
         return items
 
     def __iter__(self):
-        # Fill cache if this is the first iteration. Return an iterator over the cached results.
-        if self._cache is None:
-            log.debug('Filling cache')
-            if self.q is None:
-                self._cache = []
-            else:
-                # TODO: This is still eager processing
-                self._cache = list(self._query())
-        cache_iter = iter(self._cache)
-        return {
-            self.VALUES: lambda: self.as_values(cache_iter),
-            self.VALUES_LIST: lambda: self.as_values_list(cache_iter),
-            self.FLAT: lambda: self.as_flat_values_list(cache_iter),
-            self.NONE: lambda: cache_iter,
-        }[self.return_format]()
+        # Fill cache if this is the first iteration. Return an iterator over the results. Make this non-greedy by
+        # filling the cache while we are iterating.
+        #
+        # We don't set self._cache until the iterator is finished. Otherwise an interrupted iterator would leave the
+        # cache in an inconsistent state.
+        if self._cache is not None:
+            for val in self._cache:
+                yield val
+            return
+
+        log.debug('Initializing cache')
+        if self.q is None:
+            self._cache = []
+            return
+
+        _cache = []
+        result_formatter = {
+            self.VALUES: self.as_values,
+            self.VALUES_LIST: self.as_values_list,
+            self.FLAT: self.as_flat_values_list,
+            self.NONE: lambda res_iter: res_iter,
+        }[self.return_format]
+        for val in result_formatter(self._query()):
+            _cache.append(val)
+            yield val
+        self._cache = _cache
 
     def __len__(self):
         if self._cache is not None:
@@ -159,7 +174,7 @@ class QuerySet(object):
 
     def __getitem__(self, key):
         # Support indexing and slicing
-        self.__iter__()  # Make sure cache is full
+        list(self.__iter__())  # Make sure cache is full by iterating the full query result
         return self._cache[key]
 
     def as_values(self, iterable):
@@ -232,28 +247,32 @@ class QuerySet(object):
     # foo_qs.filter(foo='baz')  # Should not be affected by the previous statement
     #
     def all(self):
-        # Invalidate cache and return all objects
+        """ Return everything, without restrictions """
         new_qs = self.copy()
         return new_qs
 
     def none(self):
+        """ Return a query that is quaranteed to be empty  """
         new_qs = self.copy()
         new_qs.q = None
         return new_qs
 
     def filter(self, *args, **kwargs):
+        """ Return everything that matches these search criteria """
         new_qs = self.copy()
         q = Q.from_filter_args(self.folder.__class__, *args, **kwargs) or Q()
         new_qs.q = q if new_qs.q is None else new_qs.q & q
         return new_qs
 
     def exclude(self, *args, **kwargs):
+        """ Return everything that does NOT match these search criteria """
         new_qs = self.copy()
         q = ~Q.from_filter_args(self.folder.__class__, *args, **kwargs) or Q()
         new_qs.q = q if new_qs.q is None else new_qs.q & q
         return new_qs
 
     def only(self, *args):
+        """ Fetch only the specified field names. All other item fields will be 'None' """
         try:
             self._check_fields(args)
         except ValueError as e:
@@ -263,6 +282,8 @@ class QuerySet(object):
         return new_qs
 
     def order_by(self, *args):
+        """ Return the query result sorted by the specified field names. Field names prefixed with '-' will be sorted
+        in reverse order. This will make the query greedy """
         try:
             self._check_fields(f.lstrip('-') for f in args)
         except ValueError as e:
@@ -272,6 +293,7 @@ class QuerySet(object):
         return new_qs
 
     def reverse(self):
+        """ Return the entire query result in reverse order. This will make the query greedy """
         new_qs = self.copy()
         if not self.order_fields:
             raise ValueError('Reversing only makes sense if there are order_by fields')
@@ -279,6 +301,7 @@ class QuerySet(object):
         return new_qs
 
     def values(self, *args):
+        """ Return the values of the specified field names as dicts """
         try:
             self._check_fields(args)
         except ValueError as e:
@@ -289,7 +312,10 @@ class QuerySet(object):
         return new_qs
 
     def values_list(self, *args, **kwargs):
-        # Allow an arbitrary list of fields in *args, possibly ending with flat=True|False
+        """ Return the values of the specified field names as lists. If called with flat=True and only one field name,
+        return only this value instead of a list.
+
+        Allow an arbitrary list of fields in *args, possibly ending with flat=True|False"""
         flat = kwargs.pop('flat', False)
         if kwargs:
             raise AttributeError('Unknown kwargs: %s' % kwargs)
@@ -308,18 +334,20 @@ class QuerySet(object):
     #
     ###########################
     def iterator(self):
+        """ Return the query result as an iterator, without caching the result """
         if self._cache is not None:
             return self._cache
         # Return an iterator that doesn't bother with caching
         return self._query()
 
     def get(self, *args, **kwargs):
+        """ Assume the query will return exactly one item. Return that item """
         if self._cache is not None and not args and not kwargs:
-            # We cn only safely use the cache if get() is called without args
+            # We can only safely use the cache if get() is called without args
             items = self._cache
         else:
             new_qs = self.filter(*args, **kwargs)
-            items = list(new_qs)
+            items = list(new_qs.__iter__())
         if len(items) == 0:
             raise DoesNotExist()
         if len(items) != 1:
@@ -327,7 +355,7 @@ class QuerySet(object):
         return items[0]
 
     def count(self):
-        # Get the item count with as little effort as possible
+        """ Get the query count, with as little effort as possible """
         if self._cache is not None:
             return len(self._cache)
         new_qs = self.copy()
@@ -335,13 +363,14 @@ class QuerySet(object):
         new_qs.order_fields = None
         new_qs.reversed = False
         new_qs.return_format = self.NONE
-        return len(list(new_qs))
+        return len(list(new_qs.__iter__()))
 
     def exists(self):
+        """ Find out if the query contains any hits, with as little effort as possible """
         return self.count() > 0
 
     def delete(self):
-        # Delete the items with as little effort as possible
+        """ Delete the items matching the query, with as little effort as possible """
         from .folders import ALL_OCCURRENCIES
         if self._cache is not None:
             return self.folder.account.bulk_delete(ids=self._cache, affected_task_occurrences=ALL_OCCURRENCIES)